requirements.txt
awscli-local
psycopg
pycryptodome